+++ projects/ngx-reactive-loading/src/lib/model/loading-store-options.ts
@@ -2,5 +2,5 @@ export interface LoadingStoreOptions {
   /**
    * Isolate the service instance, removing the accesses to the parent and root loading state
    */
-  standalone: boolean;
+  standalone?: boolean;
 }