+++ projects/ngx-reactive-loading/src/lib/operators/index.ts
@@ -1,2 +1,1 @@
 export * from './with-loading';
-export * from './some-loading';