+++ projects/ngx-reactive-loading/src/lib/services/logger.service.ts
@@ -1,0 +1,10 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({ providedIn: 'root' })
+export class LoggerService {
+  log(title: string, content: unknown): void {
+    console.group(title);
+    console.log(content);
+    console.groupEnd();
+  }
+}