+++ projects/ngx-reactive-loading/src/spec/services/loading.service.spec.ts
@@ -1,6 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
-import { fakeAsync, TestBed, tick } from '@angular/core/testing';
-import { LoadingEvent, someLoading } from 'ngx-reactive-loading';
+import { fakeAsync } from '@angular/core/testing';
 import { combineLatest, of, timer } from 'rxjs';
 import { delay } from 'rxjs/operators';
 import { LoadingService } from '../../public-api';
@@ -22,33 +20,53 @@ describe('LoadingService', () => {
     expect(Object.keys(service.state)).toEqual(expected);
   });
 
-  it('isLoading', fakeAsync(() => {
-    const spy = jasmine.createSpy();
+  it('should throw key cannot be duplicated', () => {
+    expect(() => {
+      service = new LoadingService(
+        ['prop1', 'prop1'],
+        { standalone: true },
+        null
+      );
+    }).toThrowError(`Key 'prop1' cannot be duplicated in the current service.`);
+  });
 
-    service.load(timer(1000), 'prop1').subscribe();
+  it('should throw key already defined', () => {
+    const parent = new LoadingService(
+      ['key1', 'key2'],
+      { standalone: false },
+      null
+    );
 
-    service.isLoading('prop1').subscribe(spy);
+    expect(() => {
+      service = new LoadingService(
+        ['prop1', 'key1', 'prop2'],
+        { standalone: false },
+        parent
+      );
+    }).toThrowError(`Key 'key1' is already defined by a parent service.`);
+  });
 
-    expect(spy).toHaveBeenCalledWith(true);
+  it(
+    'isLoading',
+    marbles(m => {
+      const obs$ = service.load(timer(5), 'prop1');
+      m.expect(obs$).toBeObservable('-----(a|)', { a: 0 });
 
-    tick(1000);
+      const isLoading$ = service.isLoading('prop1');
+      m.expect(isLoading$).toBeObservable('a----b', { a: true, b: false });
+    })
+  );
 
-    expect(spy).toHaveBeenCalledWith(false);
-  }));
+  it(
+    'track',
+    marbles(m => {
+      const obs$ = timer(5).pipe(service.track('prop1'));
+      m.expect(obs$).toBeObservable('-----(a|)', { a: 0 });
 
-  it('track', fakeAsync(() => {
-    const spy = jasmine.createSpy();
-
-    timer(1000).pipe(service.track('prop1')).subscribe();
-
-    service.isLoading('prop1').subscribe(spy);
-
-    expect(spy).toHaveBeenCalledWith(true);
-
-    tick(1000);
-
-    expect(spy).toHaveBeenCalledWith(false);
-  }));
+      const isLoading$ = service.isLoading('prop1');
+      m.expect(isLoading$).toBeObservable('a----b', { a: true, b: false });
+    })
+  );
 
   it('should throw key not found on track', fakeAsync(() => {
     expect(() => {
@@ -56,78 +74,76 @@ describe('LoadingService', () => {
     }).toThrowError(Error);
   }));
 
-  it('someLoading by props', fakeAsync(() => {
-    const spy = jasmine.createSpy();
+  it(
+    'someLoading',
+    marbles(m => {
+      const prop1$ = service.load(timer(3), 'prop1');
+      m.expect(prop1$).toBeObservable('---(a|)', { a: 0 });
+      const prop2$ = service.load(timer(6), 'prop2');
+      m.expect(prop2$).toBeObservable('------(a|)', { a: 0 });
 
-    combineLatest([
-      service.load(timer(1000), 'prop1'),
-      service.load(timer(2000), 'prop2'),
-    ]).subscribe();
+      const prop1And2$ = combineLatest([prop1$, prop2$]);
+      m.expect(prop1And2$).toBeObservable('------(c|)', { c: [0, 0] });
 
-    service.someLoading(['prop1', 'prop2']).subscribe(spy);
+      const someLoading$ = service.someLoading(['prop1', 'prop2']);
+      const someLoadingByAll$ = service.someLoading();
+      const someLoadingProp1$ = service.someLoading(['prop1']);
+      const someLoadingProp2$ = service.someLoading(['prop2']);
 
-    expect(spy).toHaveBeenCalledWith(true);
+      m.expect(someLoading$).toBeObservable('a-----b', { a: true, b: false });
+      m.expect(someLoadingByAll$).toBeObservable('a-----b', {
+        a: true,
+        b: false,
+      });
+      m.expect(someLoadingProp1$).toBeObservable('a--b', { a: true, b: false });
+      m.expect(someLoadingProp2$).toBeObservable('a-----b', {
+        a: true,
+        b: false,
+      });
+    })
+  );
 
-    tick(1000);
+  it(
+    'should throw key not found on someLoading',
+    marbles(m => {
+      const notFoundProperty = 'property';
 
-    expect(spy).toHaveBeenCalledWith(true);
+      const someLoading$ = service.someLoading([notFoundProperty]);
 
-    tick(2000);
+      m.expect(someLoading$).toBeObservable(
+        '#',
+        {},
+        new Error(`[LoadingService] Property ${notFoundProperty} not found`)
+      );
+    })
+  );
 
-    expect(spy).toHaveBeenCalledWith(false);
-  }));
+  it(
+    'someLoading with parent',
+    marbles(m => {
+      const parentService = new LoadingService(
+        ['parent1'],
+        { standalone: true },
+        null
+      );
+      service = new LoadingService(
+        ['child1'],
+        { standalone: false },
+        parentService
+      );
 
-  it('someLoading by all props', fakeAsync(() => {
-    const spy = jasmine.createSpy();
+      const parent1$ = parentService.load(timer(3), 'parent1');
+      m.expect(parent1$).toBeObservable('---(a|)', { a: 0 });
+      const child1$ = service.load(timer(6), 'child1');
+      m.expect(child1$).toBeObservable('------(a|)', { a: 0 });
 
-    combineLatest([
-      service.load(timer(1000), 'prop1'),
-      service.load(timer(2000), 'prop2'),
-    ]).subscribe();
-
-    service.someLoading().subscribe(spy);
-
-    expect(spy).toHaveBeenCalledWith(true);
-
-    tick(1000);
-
-    expect(spy).toHaveBeenCalledWith(true);
-
-    tick(2000);
-
-    expect(spy).toHaveBeenCalledWith(false);
-  }));
-
-  it('someLoading with parent', fakeAsync(() => {
-    const spy = jasmine.createSpy();
-    const parentService = new LoadingService(
-      ['parent1'],
-      { standalone: true },
-      null
-    );
-    service = new LoadingService(
-      ['child1'],
-      { standalone: false },
-      parentService
-    );
-
-    combineLatest([
-      parentService.load(timer(1000), 'parent1'),
-      service.load(timer(2000), 'child1'),
-    ]).subscribe();
-
-    service.someLoading(['parent1']).subscribe(spy);
-
-    expect(spy).toHaveBeenCalledWith(true);
-
-    tick(1000);
-
-    expect(spy).toHaveBeenCalledWith(true);
-
-    tick(2000);
-
-    expect(spy).toHaveBeenCalledWith(false);
-  }));
+      const parent1SomeLoading$ = service.someLoading(['parent1']);
+      m.expect(parent1SomeLoading$).toBeObservable('a--b', {
+        a: true,
+        b: false,
+      });
+    })
+  );
 
   it(
     'would emit a new value for each property changed',